Include <stddef.h> instead of <stdlib.h> in src/wbuf.h
We just need NULL and size_t, not the whole stdlib.

+++ src/wbuf.h
@@ -1,10 +1,10 @@
 #ifndef WBUF_H
 #define WBUF_H
 
-#include <stdlib.h>
+#include <stddef.h>
 
 typedef struct {
-    int fill;
+    size_t fill;
     int fd;
     char buf[8192];
 } WriteBuffer;
